+++ AutoGossip-Core/AutoGossip-Core.lua
@@ -1,7 +1,9 @@
 local _, AutoGossip = ...
 
+_G.AutoGossip = AutoGossip
+
 AutoGossip.print_branded = function(message)
-	print("[|cffffff77Auto|r|cff00c0ffGossip|r]", message)
+	print("[|T3193419:16|t|cffffff77Auto|r|cff00c0ffGossip|r]", message)
 end
 
 AutoGossip.merge_tables = function(target, source)
@@ -21,16 +23,40 @@ AutoGossip.GOSSIPS = AutoGossip.GOSSIPS or {}
 
 AutoGossip.main = CreateFrame("Frame")
 
-function AutoGossip:OnEvent(_, type)
+function AutoGossip.main:ADDON_LOADED(addonName)
+	if addonName == "AutoGossip-Core" then
+		if AutoGossip_SeenModularizationDisclaimer == nil then
+			AutoGossip_SeenModularizationDisclaimer = false
+		end
+
+		if AutoGossip_SeenModularizationDisclaimer == false then
+			StaticPopupDialogs["MODULARIZATION_DISCLAIMER"] = {
+				text = "Starting at version 2024.09.02 of AutoGossip, automation for The War Within content has been split into a \"plugin\" addon that is included automatically. If you do not want NPCs to be automated in The War Within then you may disable the plugin in your addons menu.\n\nThis message will only be shown once.",
+				showAlert = true,
+				button1 = OKAY,
+				timeout = 0,
+				whileDead = true,
+				hideOnEscape = false
+			}
+
+			StaticPopup_Show("MODULARIZATION_DISCLAIMER")
+
+			AutoGossip_SeenModularizationDisclaimer = true
+		end
+	end
+end
+
+function AutoGossip.main:PLAYER_INTERACTION_MANAGER_FRAME_SHOW(event)
 	local should_halt = IsShiftKeyDown()
 	local should_force = IsAltKeyDown()
+	local is_gossip = event == Enum.PlayerInteractionType.Gossip
 	local has_available_quests = C_GossipInfo.GetNumAvailableQuests() > 0
 
 	----------------------------------------------------------------
 	-- Stop if SHIFT is held OR the interaction type is not Gossip
 	----------------------------------------------------------------
 	if should_halt then return end
-	if type ~= Enum.PlayerInteractionType.Gossip then return end
+	if not is_gossip then return end
 
 	----------------------------------------------------------------
 	-- Stop if we can't find the NPC name in the gossip table or
@@ -46,25 +72,25 @@ function AutoGossip:OnEvent(_, type)
 	if not options then return end
 
 	----------------------------------------------------------------
-	-- Warn if there are available quests
-	----------------------------------------------------------------
-	if has_available_quests then
-		if not should_force then
-			AutoGossip.print_branded("The NPC you are talking to has available quests. Hold |cffffff77ALT|r when speaking to them to forcefully pick their gossip option.")
-			return
-		end
-	end
-
-	----------------------------------------------------------------
 	-- Time to choose...
 	----------------------------------------------------------------
 	for _, option in ipairs(options) do
 		if tContains(gossips, option.gossipOptionID) then
-			C_GossipInfo.SelectOption(option.gossipOptionID)
-			break
+			-- Warn if there are available quests
+			if has_available_quests then
+				if not should_force then
+					AutoGossip.print_branded("The NPC you are talking to has available quests. Hold |cffffff77ALT|r when speaking to them to forcefully pick their gossip option.")
+					break
+				end
+			else
+				C_GossipInfo.SelectOption(option.gossipOptionID)
+			end
 		end
 	end
 end
 
-AutoGossip.main:SetScript("OnEvent", AutoGossip.OnEvent)
+AutoGossip.main:RegisterEvent("ADDON_LOADED")
 AutoGossip.main:RegisterEvent("PLAYER_INTERACTION_MANAGER_FRAME_SHOW")
+AutoGossip.main:SetScript("OnEvent", function(self, event, ...)
+	self[event](self, ...)
+end)