+++ build.ps1
@@ -1,13 +1,20 @@
 $scriptDir = Split-Path -Parent $MyInvocation.MyCommand.Definition
 $filesToZip = @(
-	"$scriptDir\AutoGossip.lua",
-	"$scriptDir\AutoGossip.toc",
-	"$scriptDir\gossips\Instances.lua",
-	"$scriptDir\gossips\OpenWorld.lua",
-	"$scriptDir\gossips\Toys.lua",
-	"$scriptDir\gossips\Vendors.lua",
+	# Core
+	"$scriptDir\AutoGossip-Core\AutoGossip-Core.lua",
+	"$scriptDir\AutoGossip-Core\AutoGossip-Core.toc",
+	"$scriptDir\AutoGossip-Core\gossips\Instances.lua",
+	"$scriptDir\AutoGossip-Core\gossips\OpenWorld.lua",
+	"$scriptDir\AutoGossip-Core\gossips\Toys.lua",
+	"$scriptDir\AutoGossip-Core\gossips\Vendors.lua",
+	# TWW Plugin
+	"$scriptDir\AutoGossip-TWW\AutoGossip-TWW.lua",
+	"$scriptDir\AutoGossip-TWW\AutoGossip-TWW.toc"
+)
+$alwaysCopy = @(
 	"$scriptDir\CHANGELOG.md",
-	"$scriptDir\LICENSE"
+	"$scriptDir\LICENSE",
+	"$scriptDir\README.md"
 )
 $destinationFolder = "$scriptDir\builds"
 $zipFolderName = "AutoGossip"
@@ -19,7 +26,8 @@ $counter = 1
 function Create-Zip {
 	param (
 		[string]$zipFileName,
-		[string[]]$files
+		[string[]]$files,
+		[string[]]$extraFiles
 	)
 	Add-Type -AssemblyName "System.IO.Compression.FileSystem"
 
@@ -36,6 +44,13 @@ function Create-Zip {
 		Copy-Item -Path $file -Destination $destinationPath -Force
 	}
 
+	$subFolders = Get-ChildItem -Directory -Path $folderInZip
+	foreach ($subFolder in $subFolders) {
+		foreach ($extraFile in $extraFiles) {
+			Copy-Item -Path $extraFile -Destination $subFolder.FullName -Force
+		}
+	}
+
 	[System.IO.Compression.ZipFile]::CreateFromDirectory($tempFolder, $zipFileName)
 
 	Remove-Item -Path $tempFolder -Recurse -Force
@@ -46,6 +61,6 @@ while (Test-Path $zipFileName) {
 	$counter++
 }
 
-Create-Zip -zipFileName $zipFileName -files $filesToZip
+Create-Zip -zipFileName $zipFileName -files $filesToZip -extraFiles $alwaysCopy
 
 Write-Output "Created build: $zipFileName"